responses array added
