+++ reg-portal/src/components/Quiz/TechQuiz.js
@@ -5,18 +5,17 @@ import Background from "../../hoc/Background/Background";
 import { Redirect } from "react-router-dom";
 
 class TechQuiz extends React.Component {
+    selectedOptions= [];
     constructor(props) {
         super(props);
         this.state =  {
             quizQuestions: [],
             time: 600,
             currentQuestionIndex: 0,
-            selectedOptions: [],
             questionId: []
         };
         this.setSelectedOption = this.setSelectedOption.bind(this);
     }
-
     getTimer() {
         if (this.state.time > 0) {
             setTimeout(() => {
@@ -52,15 +51,19 @@ class TechQuiz extends React.Component {
 
     optionsArray = ["a", "b", "c", "d"];
 
-    setSelectedOption() {
-        console.log("Hi");
-        this.setState({
-            selectedOptions: [...this.state.selectedOptions, this.optionsArray[1]]
-
-            // const selectedOptions = state.selectedOptions.concat(this.optionsArray[index]);
-            // const questionId = state.questionId.concat(id);
-        })
-        console.log(this.state.selectedOptions);
+    setSelectedOption(id,index) {
+        // console.log(id,this.optionsArray[index]);
+        if(this.selectedOptions.some(option=> option.id ===id)){
+            for(let i=0;i<this.selectedOptions.length;i++){
+                if(this.selectedOptions[i].id===id){
+                    console.log("okay match");
+                    this.selectedOptions[i].index=this.optionsArray[index];
+                }
+            }
+        } else {
+            this.selectedOptions.push({id:id,index:this.optionsArray[index]})
+        }
+        console.log("hello",this.selectedOptions);
     }
 
     gotoNextQuestion() {
@@ -114,9 +117,8 @@ class TechQuiz extends React.Component {
                             <div className='answer-section'>
                                 {Object.keys(this.state.quizQuestions[this.state.currentQuestionIndex].options).map((key, index) => {
                                     return (
-                                        <div key={index}>
+                                        <div key={index} onClick={()=>this.setSelectedOption(this.state.quizQuestions[this.state.currentQuestionIndex]._id,index)}>
                                             <button className="options"
-                                                    onCLick={ () => this.setSelectedOption() }
                                                     value={this.optionsArray[index]}>
                                                     {this.optionsArray[index]}. {this.state.quizQuestions[this.state.currentQuestionIndex].options[key]}
                                             </button>
